packaging [11/n]
fix: add an conditional branch with a nested unconditional
branch holding the logic for [REBUILDing] the OCI image when
a previously built is found. No pun intended. Or was it? :)

add: the isogen/iso9660 branch custom badge for the qonq_iptables
workflow runner

fix: adjust name of the workflow runners for qonq_iptables and
qonq_shadow to keep the consistency between contexts

fix: call to the makefile -> shellscript -> dockerfile -> shellscript
context inside the qonq_iptables workflow runner.